factoring / as library
